Detect AMD GPUs without lspci
shelling out to lspci and relying upon pciids.txt is fragile,
especially when trying to run on an old distro.  To avoid risk with
this instead parse amdgpu sysfs files to determine if an APU or dGPU
and build the graphics architecture from the HW IP version.

NOTE: this is not a perfect solution still because HW IP and SW IP
version might not match.  But it's certainly A LOT closer to right.

There might need to be a matrix to map out HW IP to SW IP versions
as another layer.

Signed-off-by: Mario Limonciello (AMD) <superm1@kernel.org>